Created Json file for node
package.json file created and along with it our node_modules folder was
created. This is where
the node modules this app depends on are installed. the json file hold
a list of the dependencies of this app, versioning, git repo and a few
things we can fill out later, like description, key words etc.
Node modules Express and Sock.io have both been installed and saved to
the json file