Coordinator Wired To Server
The coordinator page is now wired up to the node server. It enables the
judge’s voting buttons when the timer is started. When a decisions has
been made by the judges the coordinators timer is paused and the
coordinator can then click stop but also has the power to resume if an
error has been made. Also when the stop is hit the judge screens revert
back to a clone of the div made before any votes were cast and this
sets the judge up to make the next series of votes. We will have to
modify this some once it is connected to the spreadsheet.